+++ src/briefy/leica/sync/job.py
@@ -10,6 +10,8 @@ from briefy.leica.sync import ModelSync
 from briefy.leica.sync.location import create_location_dict
 from briefy.leica.vocabularies import JobInputSource as ISource
 
+import uuid
+
 
 job_status_mapping = {
     'Pending': 'pending',
@@ -88,6 +90,7 @@ class JobSync(ModelSync):
         if payload:
             payload['order_id'] = obj.id
             payload.update(
+                id=uuid.uuid4(),
                 mobile=self.parse_phonenumber(kobj, 'contact_number_1'),
                 additional_phone=self.parse_phonenumber(kobj, 'contact_number_2'),
                 email=kobj.contact_email.email or 'abc123@gmail.com',
@@ -98,7 +101,6 @@ class JobSync(ModelSync):
                 location = JobLocation(**payload)
                 self.session.add(location)
                 obj.locations.append(location)
-                self.session.flush()
             except Exception as exc:
                 print(exc)
         else:
@@ -109,6 +111,7 @@ class JobSync(ModelSync):
         """Add Project Manager comment to the Order."""
         project_manager = obj.project.project_manager if obj.project.project_manager else None
         payload = dict(
+            id=uuid.uuid4(),
             entity_id=obj.id,
             entity_type=obj.__tablename__,
             author_id=project_manager,
@@ -116,7 +119,6 @@ class JobSync(ModelSync):
         )
         session = self.session
         session.add(Comment(**payload))
-        session.flush()
 
     def add_assigment_comments(self, obj, kobj):
         """Import assigment comments"""
@@ -135,6 +137,7 @@ class JobSync(ModelSync):
 
         professional_id = self.get_user(kobj, 'responsible_photographer')
         payload = dict(
+            id=uuid.uuid4(),
             order_id=obj.id,
             professional_id=professional_id,
             payout_value=self.parse_decimal(kobj.photographer_payout),
@@ -146,7 +149,6 @@ class JobSync(ModelSync):
         )
         assignment = JobAssignment(**payload)
         self.session.add(assignment)
-        self.session.flush()
         logger.debug('Assignment added: {id}'.format(id=assignment.id))
 
         if professional_id:
@@ -191,5 +193,5 @@ class JobSync(ModelSync):
         self.add_location(obj, kobj)
         self.add_comment(obj, kobj)
         self.add_assigment(obj, kobj)
-        logger.info('Additional data imported for this order: History, Location, Comment.')
+        logger.debug('Additional data imported for this order: History, Location, Comment.')
         return obj