+++ src/briefy/leica/tools/customer_import.py
@@ -4,15 +4,12 @@ from briefy.leica.sync.customer import CustomerSync
 from briefy.leica.tools import logger # noqa
 
 
-import transaction
-
-
-def main(session):
+def main(session, trasaction):
     """Import Customer script."""
-    CustomerSync(session)()
+    CustomerSync(session, transaction)()
 
 
 if __name__ == '__main__':
     session = configure(Session)
-    with transaction.manager:
-        main(session)
+    import transaction
+    main(session, transaction)